requirements.txt
aiohttp==3.13.3
discord.py==2.7.1
